app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["parkhongbeen <pack122@naver.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django-rest-framework = "^0.1.0"
drf-yasg = "^1.17.1"
drf-writable-nested = "^0.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
